# src/main/kotlin/com/signicat/interview/gen/api/UsersApiDelegate.kt
package com.signicat.interview.gen.api

import com.signicat.interview.gen.api.model.InlineObjectDto
import com.signicat.interview.gen.api.model.UserDto
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.core.io.Resource

import java.util.Optional

/**
 * A delegate to be called by the {@link UsersApiController}}.
 * Implement this interface with a {@link org.springframework.stereotype.Service} annotated class.
 */
@javax.annotation.Generated(value = ["org.openapitools.codegen.languages.KotlinSpringServerCodegen"])
interface UsersApiDelegate {

    fun getRequest(): Optional<NativeWebRequest> = Optional.empty()

    /**
     * @see UsersApi#getTest
     */
    fun getTest(): ResponseEntity<kotlin.String> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)

    }


    /**
     * @see UsersApi#login
     */
    fun login(inlineObjectDto: InlineObjectDto): ResponseEntity<kotlin.String> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)

    }


    /**
     * @see UsersApi#registerUser
     */
    fun registerUser(userDto: UserDto): ResponseEntity<UserDto> {
        getRequest().ifPresent { request ->
            for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    ApiUtil.setExampleResponse(request, "application/json", "{  \"password\" : \"password\",  \"userGroupNames\" : [ \"userGroupNames\", \"userGroupNames\" ],  \"id\" : 0,  \"username\" : \"username\"}")
                    break
                }
            }
        }
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)

    }

}